Changing the Direction.
Managing frontend is harder than I expected. Thus I will stick to my structure of backend code with frontend elements, interchangeable to the degree I can handle. I really want someone experienced on my side to guide me. I feel like my Project is not as easy to expand as it could be. I'll probably refactor it to use Spring Boot in the future. For now, I resign from using JavaFX as a whole, only some of it's components.